+++ tests/directory_watcher/test_directory_watcher.py
@@ -22,6 +22,7 @@ class TestDirectoryWatcher:
 
     STORAGE_NAME = "data"
     INGREST_SERVER_URL = "http://localhost:8001"
+    ROOT_DIRECTORY = "/data"
 
     add_path_successful = False
 
@@ -38,6 +39,8 @@ class TestDirectoryWatcher:
                 cls.INGREST_SERVER_URL,
                 "--storage-name",
                 cls.STORAGE_NAME,
+                "--storage-root-directory",
+                cls.ROOT_DIRECTORY,
             ]
         )
         cls.config = process_args(args=cls.parsed)
@@ -92,9 +95,11 @@ class TestDirectoryWatcher:
         a_temp_file_relative_path = a_temp_file.replace(f"{self.the_watch_dir}/", "")
         # On MacOS the system messes with the path by adding a /private
         absolute_path = registration_processor.absolute_path.replace("/private", "")
-        relative_path = registration_processor.relative_path.replace("/private", "")
+        path_rel_to_watch_dir = registration_processor.path_rel_to_watch_dir.replace(
+            "/private", ""
+        )
         assert a_temp_file == absolute_path
-        assert a_temp_file_relative_path == relative_path
+        assert a_temp_file_relative_path == path_rel_to_watch_dir
         Path(a_temp_file).unlink()
 
 
@@ -102,9 +107,9 @@ class MockRegistrationProcessor(RegistrationProcessor):
     """A class to use for test of directory watcher."""
 
     absolute_path: str
-    relative_path: str
+    path_rel_to_watch_dir: str
 
-    def add_path(self, absolute_path: str, relative_path: str):
+    def add_path(self, absolute_path: str, path_rel_to_watch_dir: str):
         """Perform testing on the given paths."""
         self.absolute_path = absolute_path
-        self.relative_path = relative_path
+        self.path_rel_to_watch_dir = path_rel_to_watch_dir